2-level transaction multi-signatures (RIPD-182):
This commit provides support for 2-level multi-signing of
transactions.  The ability is usually compiled out, since other
aspects of multi-signing are not yet complete.

Here are the missing parts:

 o Full support for Tickets in transactions.
 o Variable fees based on the number of signers,
 o Multiple SignerLists with access control flags on accounts,
 o Enable / disable operations based on access control flags,
 o Enable / disable all of multi-signing based on an amendment,
 o Integration tests, and
 o Documentation.

+++ src/ripple/app/ledger/LedgerEntrySet.h
@@ -185,13 +185,13 @@ public:
     /** @{ */
     void incrementOwnerCount (SLE::ref sleAccount);
     void incrementOwnerCount (Account const& owner);
-    void increaseOwnerCount (SLE::ref sleAccount, std::size_t howMuch);
+    void increaseOwnerCount (SLE::ref sleAccount, std::uint32_t howMuch);
     /** @} */
 
     /** @{ */
     void decrementOwnerCount (SLE::ref sleAccount);
     void decrementOwnerCount (Account const& owner);
-    void decreaseOwnerCount (SLE::ref sleAccount, std::size_t howMuch);
+    void decreaseOwnerCount (SLE::ref sleAccount, std::uint32_t howMuch);
     /** @} */
 
     // Offer functions.
